# api/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: API server
paths:
  /v1/registration:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Bad request - Invalid JSON or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity - Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/login:
    post:
      summary: Login and get authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request - Invalid JSON or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity - Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/profile:
    get:
      summary: Get authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - User does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tasks:
    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTaskResponse'
        '400':
          description: Bad request - Invalid JSON or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity - Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - passwordConfirmation
      properties:
        email:
          type: string
          example: hello@world.com
        password:
          type: string
          format: password
          example: SecurePassword123!
        passwordConfirmation:
          type: string
          format: password
          example: SecurePassword123!
    User:
      type: object
      required:
        - email
        - created_at
      properties:
        email:
          type: string
          example: hello@world.com
        created_at:
          type: string
          format: date-time
          example: "2025-11-16T10:30:00Z"
    CreateUserResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user
          properties:
            user:
              $ref: '#/components/schemas/User'
    GetUserResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user
          properties:
            user:
              $ref: '#/components/schemas/User'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: hello@world.com
        password:
          type: string
          format: password
          example: SecurePassword123!
    LoginResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user
            - token
            - expiresAt
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            expiresAt:
              type: string
              format: date-time
              example: "2025-11-17T10:30:00Z"
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          example: "Write comprehensive documentation for the API"
        deadline:
          type: string
          format: date
          example: "2025-12-31"
    Task:
      type: object
      required:
        - id
        - title
        - is_done
        - created_at
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          example: "Write comprehensive documentation for the API"
        deadline:
          type: string
          format: date
          example: "2025-12-31"
        is_done:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-11-17T10:30:00Z"
    CreateTaskResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - task
          properties:
            task:
              $ref: '#/components/schemas/Task'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              example: "Email is required"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT